Infer logical reading order from pdfminer, remove --cols argument.
This is a major change to the internal logic. Rather than hardcoding a
particular page layout for comparing x,y coordinates on a page, we now rely on
pdfminer's own text layout detection, and sort locations (annotations and
outlines) based on their proximity to text lines. Since we are already fully
dependent on pdfminer to extract text in a sensible order, this hopefully
reduces the amount of extra heuristics in pdfannots, and definitely makes things
simpler for the user.

There's also a non-trivial amount of refactoring, tidying and commenting.