restore the ability to fix the page layout via --cols parameter if desired
I discovered that PDFMiner's layout algorithm is less reliable than desired, and even for the
checked-in hotos17 test it was producing the wrong order of text leading to the wrong outlines
being chosen in markdown output. This restores the ability to use the old fixed layout order,
only if desired, by passing --cols. If not specified, the new layout-based order is used.