Handle "Square" annotations, but don't attempt to capture text
Per issue #46, Apple tools support "highlighting" PDFs where the hightlights
are emitted as a Square annotation with a custom appearance that renders the
markup. Figuring out the affected text would be a major undertaking, but
with this change we (1) recognise the existence of the annotation rather than
emitting an "unsupported annotation" warning, and (2) capture the contents
(text note) of the annotation if any.