Change Markdown output for deletions from HTML <del>foo</del> to ~~foo~~.
Classic Markdown lacks a standard format for deletions, but allows the use
of HTML tags. HotCRP doesn't support HTML tags, but does support the ~~
syntax, which is apparently quite common. Because the primary motivation of
this tool is (a) legible plain text output, and (b) HotCRP reviews, this
feels like a better choice.