feat(markdown): Add defaultdict for grouped highlights
This commit adds a defaultdict to hold grouped highlights by color in the GroupedMarkdownPrinter class. If the highlights need to be grouped by color, they will be stored in the defaultdict with the color as the key. Otherwise, they will be stored in a normal list. This change improves efficiency and readability when printing annotations in markdown format.